Método de adicionar uma encomenda

+++ src/main/java/rw/solutions/api/resi/model/record/DadosEncomenda.java
@@ -3,12 +3,12 @@ package rw.solutions.api.resi.model.record;
 import rw.solutions.api.resi.model.Encomenda;
 import rw.solutions.api.resi.model.EncomendaStatus;
 
-public record DadosEncomenda(int id,
+public record DadosEncomenda(Long id,
 							 String nome,
 							 String assinadoPor,
 							 EncomendaStatus status,
-							 int moradorID,
-							 int apartamentoID) {
+							 Long moradorID,
+							 Long apartamentoID) {
 	
 	
 	public DadosEncomenda(Encomenda encomenda) {
